Moved shouldFormBeSent() to InstanceSubmitter
